backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
python_multipart
python_jose
jwt
psycopg2
pytz
apscheduler